casting long to downrate.

+++ src/project/Server.java
@@ -187,7 +187,7 @@ public class Server extends Thread{
 										Long downtime = endTime - neighbor.startTime.get(neighbor.peerId);
 										System.out.println("downtime here **** -> "+downtime);
 										System.out.println("pieceSize here **** -> "+pieceSize);
-										Peer.getInstance().downloadTime.put(neighbor.peerId, (pieceSize)/downtime);
+										Peer.getInstance().downloadTime.put(neighbor.peerId, (long)(pieceSize)/downtime);
 									}
 									pieceSize = 0;
 									Peer.getInstance().getBitField().set(reqPieceInd);
